Refactor records to remove duplicate code
