+++ src/reconcilers/bind9instance/cluster_helpers_tests.rs
@@ -1,0 +1,59 @@
+// Copyright (c) 2025 Erick Bourgeois, firestoned
+// SPDX-License-Identifier: MIT
+
+//! Unit tests for `cluster_helpers.rs`
+//!
+//! These tests document expected behavior for cluster integration helpers.
+//! Full implementation requires Kubernetes API mocking infrastructure.
+
+#[cfg(test)]
+mod tests {
+    #[tokio::test]
+    async fn test_fetch_cluster_info_from_owner_reference() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: An instance with ownerReferences containing a `Bind9Cluster`
+        //        AND the cluster exists in the same namespace
+        // When: fetch_cluster_info is called
+        // Then: Should return (Some(cluster), None)
+        //       AND log "Found cluster from ownerReference: Bind9Cluster/<name>"
+    }
+
+    #[tokio::test]
+    async fn test_fetch_cluster_info_from_provider_owner() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: An instance with ownerReferences containing a `ClusterBind9Provider`
+        //        AND the provider exists (cluster-scoped)
+        // When: fetch_cluster_info is called
+        // Then: Should return (None, Some(provider))
+        //       AND log "Found cluster from ownerReference: ClusterBind9Provider/<name>"
+    }
+
+    #[tokio::test]
+    async fn test_fetch_cluster_info_from_cluster_ref() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: An instance with spec.clusterRef = "my-cluster"
+        //        AND no ownerReferences
+        //        AND a `Bind9Cluster` named "my-cluster" exists in the same namespace
+        // When: fetch_cluster_info is called
+        // Then: Should return (Some(cluster), None)
+        //       AND log "Found cluster from clusterRef: Bind9Cluster/<name>"
+    }
+
+    #[tokio::test]
+    async fn test_fetch_cluster_info_no_cluster_found() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: An instance with spec.clusterRef = "missing-cluster"
+        //        AND no cluster exists with that name
+        // When: fetch_cluster_info is called
+        // Then: Should return (None, None)
+        //       AND log warning about missing cluster
+    }
+}
