+++ src/reconcilers/bind9instance/resources_tests.rs
@@ -1,0 +1,117 @@
+// Copyright (c) 2025 Erick Bourgeois, firestoned
+// SPDX-License-Identifier: MIT
+
+//! Unit tests for `resources.rs`
+//!
+//! These tests document expected behavior for Kubernetes resource management.
+//! Full implementation requires Kubernetes API mocking infrastructure.
+
+#[cfg(test)]
+mod tests {
+    #[tokio::test]
+    async fn test_reconcile_resources_creates_all() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: A new `Bind9Instance` with no existing resources
+        // When: reconcile_resources is called
+        // Then: Should create `ConfigMap`, RNDC Secret, Service, and Deployment
+        //       AND return Ok(())
+        //       AND log "Created all resources for instance"
+    }
+
+    #[tokio::test]
+    async fn test_reconcile_resources_updates_existing() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: An instance with existing resources
+        //        AND spec.version changed from "9.18.20" to "9.18.24"
+        // When: reconcile_resources is called
+        // Then: Should update Deployment with new image version
+        //       AND update `ConfigMap` if configuration changed
+        //       AND log "Updated resources for instance"
+    }
+
+    #[tokio::test]
+    async fn test_reconcile_resources_skips_unchanged() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: An instance with existing resources
+        //        AND no spec changes
+        // When: reconcile_resources is called
+        // Then: Should NOT update any resources
+        //       AND log "Resources unchanged, skipping update"
+    }
+
+    #[tokio::test]
+    async fn test_create_configmap_with_cluster_config() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: An instance referencing a cluster with global configuration
+        // When: create_configmap is called
+        // Then: Should create `ConfigMap` with cluster-level config
+        //       AND include named.conf with cluster settings
+        //       AND include named.conf.options
+    }
+
+    #[tokio::test]
+    async fn test_create_configmap_with_custom_refs() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: An instance with spec.configMapRefs set
+        // When: create_configmap is called
+        // Then: Should skip `ConfigMap` creation
+        //       AND log "Instance uses custom `ConfigMaps`, skipping creation"
+    }
+
+    #[tokio::test]
+    async fn test_create_deployment_with_replicas() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: An instance with spec.replicas = 3
+        // When: create_deployment is called
+        // Then: Should create Deployment with 3 replicas
+        //       AND set anti-affinity rules for pod distribution
+    }
+
+    #[tokio::test]
+    async fn test_create_service_primary_with_loadbalancer() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: A primary instance
+        // When: create_service is called
+        // Then: Should create Service with type=LoadBalancer
+        //       AND expose DNS ports (53/TCP, 53/UDP, 953/TCP)
+    }
+
+    #[tokio::test]
+    async fn test_create_service_secondary_cluster_ip() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: A secondary instance
+        // When: create_service is called
+        // Then: Should create Service with type=ClusterIP
+        //       AND expose DNS ports (53/TCP, 53/UDP, 953/TCP)
+    }
+
+    #[tokio::test]
+    async fn test_delete_resources_cleanup() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: An instance being deleted (has deletionTimestamp)
+        // When: delete_resources is called
+        // Then: Should delete Deployment first (graceful shutdown)
+        //       AND delete Service
+        //       AND delete `ConfigMap`
+        //       AND delete RNDC Secret
+        //       AND log "Successfully deleted all resources"
+    }
+}
