+++ src/reconcilers/bind9instance/status_helpers_tests.rs
@@ -1,0 +1,96 @@
+// Copyright (c) 2025 Erick Bourgeois, firestoned
+// SPDX-License-Identifier: MIT
+
+//! Unit tests for `status_helpers.rs`
+//!
+//! These tests document expected behavior for status calculation and updates.
+//! Full implementation requires Kubernetes API mocking infrastructure.
+
+#[cfg(test)]
+mod tests {
+    #[tokio::test]
+    async fn test_calculate_status_deployment_ready() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: An instance with Deployment status Ready=True
+        //        AND all pods running
+        //        AND replicas = 3, ready replicas = 3
+        // When: calculate_status is called
+        // Then: Should return Ready=True condition
+        //       AND message "All 3 pods are ready"
+        //       AND reason "DeploymentReady"
+    }
+
+    #[tokio::test]
+    async fn test_calculate_status_deployment_not_ready() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: An instance with Deployment status Ready=False
+        //        AND replicas = 3, ready replicas = 1
+        // When: calculate_status is called
+        // Then: Should return Ready=False condition
+        //       AND message "1/3 pods are ready"
+        //       AND reason "DeploymentNotReady"
+    }
+
+    #[tokio::test]
+    async fn test_calculate_status_deployment_progressing() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: An instance with Deployment in Progressing state
+        //        AND replicas = 3, ready replicas = 2
+        // When: calculate_status is called
+        // Then: Should return Ready=False condition
+        //       AND message "Deployment progressing: 2/3 pods ready"
+        //       AND reason "DeploymentProgressing"
+    }
+
+    #[tokio::test]
+    async fn test_calculate_status_no_deployment() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: An instance with no Deployment found
+        // When: calculate_status is called
+        // Then: Should return Ready=False condition
+        //       AND message "Deployment not found"
+        //       AND reason "DeploymentMissing"
+    }
+
+    #[tokio::test]
+    async fn test_update_status_patches_when_changed() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: An instance with current status Ready=False
+        //        AND new calculated status Ready=True
+        // When: update_status is called
+        // Then: Should patch the instance status
+        //       AND log "Updated instance status to Ready=True"
+    }
+
+    #[tokio::test]
+    async fn test_update_status_skips_when_unchanged() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: An instance with current status matching new status
+        // When: update_status is called
+        // Then: Should NOT patch the instance
+        //       AND log "Status unchanged, skipping update"
+    }
+
+    #[tokio::test]
+    async fn test_update_status_sets_observed_generation() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: An instance with metadata.generation = 5
+        // When: update_status is called
+        // Then: Should set status.observedGeneration = 5
+        //       AND include in the status patch
+    }
+}
