+++ src/reconcilers/bind9instance/zones_tests.rs
@@ -1,0 +1,95 @@
+// Copyright (c) 2025 Erick Bourgeois, firestoned
+// SPDX-License-Identifier: MIT
+
+//! Unit tests for `zones.rs`
+//!
+//! These tests document expected behavior for zone reference reconciliation.
+//! Full implementation requires Kubernetes API mocking infrastructure.
+
+#[cfg(test)]
+mod tests {
+    #[tokio::test]
+    async fn test_reconcile_zone_references_no_zones() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: An instance that is not selected by any `DNSZone` resources
+        // When: reconcile_zone_references is called
+        // Then: Should set status.zones = []
+        //       AND log "Instance not selected by any zones"
+    }
+
+    #[tokio::test]
+    async fn test_reconcile_zone_references_multiple_zones() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: An instance selected by 3 `DNSZone` resources
+        //        AND zones have names: "example.com", "internal.local", "test.zone"
+        // When: reconcile_zone_references is called
+        // Then: Should set status.zones with 3 zone references
+        //       AND each reference includes zone name and namespace
+        //       AND log "Found 3 zones selecting this instance"
+    }
+
+    #[tokio::test]
+    async fn test_reconcile_zone_references_zone_removed() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: An instance with status.zones = ["zone-a", "zone-b"]
+        //        AND zone-b no longer selects this instance
+        // When: reconcile_zone_references is called
+        // Then: Should update status.zones = ["zone-a"]
+        //       AND log "Zone zone-b no longer selects this instance"
+    }
+
+    #[tokio::test]
+    async fn test_reconcile_zone_references_zone_added() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: An instance with status.zones = ["zone-a"]
+        //        AND zone-b now selects this instance (via bind9InstancesFrom)
+        // When: reconcile_zone_references is called
+        // Then: Should update status.zones = ["zone-a", "zone-b"]
+        //       AND log "New zone zone-b now selects this instance"
+    }
+
+    #[tokio::test]
+    async fn test_find_zones_selecting_instance() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: 5 `DNSZone` resources in the namespace
+        //        AND 2 zones have bind9InstancesFrom selectors matching this instance
+        //        AND 3 zones do not match
+        // When: find_zones_selecting_instance is called
+        // Then: Should return vec with 2 zone references
+        //       AND include zone names and namespaces
+    }
+
+    #[tokio::test]
+    async fn test_find_zones_selecting_instance_label_match() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: An instance with labels {role: primary, env: prod}
+        //        AND a zone with bind9InstancesFrom selector {role: primary}
+        // When: find_zones_selecting_instance is called
+        // Then: Should match the zone (subset match)
+        //       AND include it in returned references
+    }
+
+    #[tokio::test]
+    async fn test_find_zones_selecting_instance_no_match() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: An instance with labels {role: primary}
+        //        AND a zone with bind9InstancesFrom selector {role: secondary}
+        // When: find_zones_selecting_instance is called
+        // Then: Should NOT match the zone
+        //       AND return empty vec
+    }
+}
